requirements.txt
gamsapi[transfer]
pandas